2 examples work
